Use type aliases to abstract the inotify code
